LayoutTests/imported/w3c/web-platform-tests/css/css-variables/variable-empty-name-reserved-expected.txt


PASS -- is a reserved property name
